refactor: configure clerk pages
