+++ .env.example
@@ -8,6 +8,7 @@ NEXT_PUBLIC_STREAMING_BASE_URL=https://097c-190-105-0-25.ngrok-free.app
 NEXT_PUBLIC_APP_CAMERA_COUNT=2
 
 # Clave publicable de Clerk
+# Reemplaza con tu clave publicable de Clerk
 NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY=
 
 # Rutas de autenticación de Clerk
