+++ src/views/auth/index.ts
@@ -1,3 +1,1 @@
-export { Login } from './Login';
-export { Register } from './Register';
-export { Welcome } from './Welcome';
+export { Welcome } from "./Welcome";
